requirements.txt
appnope==0.1.2
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
click==8.0.0
colorama==0.4.4
decorator==4.4.2
Flask==2.0.0
fuzzywuzzy==0.18.0
gitdb==4.0.7
GitPython==3.1.18
hvac==0.10.14
idna==2.10
ipython==7.19.0
ipython-genutils==0.2.0
itsdangerous==2.0.0
jedi==0.18.0
Jinja2==3.0.0
MarkupSafe==2.0.0
newrelic==6.4.4.161
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.8
ptyprocess==0.7.0
Pygments==2.7.3
pyhcl==0.4.4
python-gitlab==2.9.0
questionary==1.9.0
requests==2.25.1
requests-toolbelt==0.9.1
six==1.16.0
smmap==4.0.0
traitlets==5.0.5
urllib3==1.26.2
wcwidth==0.2.5
Werkzeug==2.0.0
